refactor: 클래스명 수정 및 적용
