+++ src/main/java/roomescape/auth/util/CookieUtil.java
@@ -1,20 +1,20 @@
-package roomescape.auth;
+package roomescape.auth.util;
 
 import org.springframework.stereotype.Service;
 
 import jakarta.servlet.http.Cookie;
 
 @Service
-public class CookieService {
+public class CookieUtil {
 
-	public Cookie createCookie(String token) {
+	public static Cookie createCookie(String token) {
 		Cookie cookie = new Cookie("token", token);
 		cookie.setHttpOnly(true);
 		cookie.setPath("/");
 		return cookie;
 	}
 
-	public String extractTokenFromCookie(Cookie[] cookies) {
+	public static String extractTokenFromCookie(Cookie[] cookies) {
 		for (Cookie cookie : cookies) {
 			if (cookie.getName().equals("token")) {
 				return cookie.getValue();
